backend/requirements.txt
# Core FastAPI Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# Configuration and Settings
pydantic>=2.4.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# AI & ML - Using latest unified SDK
google-genai>=0.6.0              # Unified SDK for both Google AI Studio & Vertex AI
google-cloud-aiplatform>=1.48.0  # Latest platform SDK for enterprise features
sentence-transformers>=2.2.0

# Database & Storage
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0                  # PostgreSQL async driver
psycopg2-binary>=2.9.0           # PostgreSQL binary
pgvector>=0.2.0                  # PostgreSQL vector extension
sqlite-vec>=0.1.0
chromadb>=0.4.0

# Document Processing for RAG
PyPDF2>=3.0.0
python-docx>=0.8.0
python-pptx>=0.6.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-magic>=0.4.0
pillow>=10.0.0
pytesseract>=0.3.0

# Text Processing
nltk>=3.8.0
spacy>=3.7.0

# Caching
redis>=5.0.0
redis[hiredis]>=5.0.0
diskcache>=5.6.0

# Background Tasks
celery>=5.3.0

# Monitoring & Logging
psutil>=5.9.0
python-json-logger>=2.0.0

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Advanced Features (Phase 2/3)
# networkx>=3.0
# spacy>=3.7.0
# langchain>=0.1.0
# faiss-cpu>=1.7.0